Python/Experiment 2 - Opus/trading_bot/requirements.txt
flask==3.1.0
gunicorn==23.0.0
python-bitvavo-api==1.4.0
python-dotenv==1.1.0
